实现新增/修改课程计划功能

+++ xuecheng-plus-content/xuecheng-plus-content-model/src/main/java/com/xuecheng/content/model/po/Teachplan.java
@@ -9,7 +9,7 @@ public class Teachplan {
     private Long id;
     private String pname;
     private Long parentid;
-    private Short grade;
+    private Integer grade;
     private String mediaType;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
